requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.10.2
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrgh==1.0.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.8.2
bokeh==2.4.3
Bottleneck==1.3.7
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
certipy==0.1.3
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cmake==3.25.0
colorama==0.4.6
comm==0.1.2
conda==23.1.0
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
Cython==0.29.33
cytoolz==0.12.0
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2023.4.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distributed==2023.3.0
dominate==2.7.0
einops==0.8.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.0
Flask==3.0.3
flatbuffers==23.1.21
flit_core==3.8.0
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
gast==0.4.0
gmpy2==2.1.2
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.0
h5py==3.8.0
HeapDict==1.0.1
hydra-core==1.3.2
idna==3.4
imagecodecs==2023.1.23
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipympl==0.9.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.2.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.5
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter-events==0.6.3
jupyter_server==2.4.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyter-telemetry==0.1.0
jupyter-ydoc==0.2.2
jupyterhub==3.1.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
jupyterlab-widgets==3.0.5
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
libmambapy==1.3.1
lightning-utilities==0.8.0
lit==15.0.7
llvmlite==0.39.1
locket==1.0.0
lz4==4.3.2
Mako==1.2.4
mamba==1.3.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.1.0
monai==1.3.1
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
munkres==1.1.4
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nibabel==5.1.0
notebook==6.5.3
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.3
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
omegaconf==2.3.0
openpyxl==3.1.1
opt-einsum==3.3.0
packaging==23.0
pamela==1.0.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
pkgutil_resolve_name==1.3.10
platformdirs==3.1.0
plotly==5.22.0
pluggy==1.0.0
pooch==1.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.4
pycparser==2.21
pycurl==7.45.1
pydot==1.4.2
Pygments==2.14.0
pyimzML==1.5.4
PyJWT==2.6.0
pynndescent==0.5.12
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.2
pytz==2022.7.1
pyu2f==0.1.5
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==67.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.5.post1
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.11.1
tables==3.7.0
tblib==1.7.0
tenacity==8.3.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.2.28
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchmetrics==0.11.4
torchvision==0.18.0+cu118
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transforms==0.1
triton==2.3.0
typing_extensions==4.9.0
ujson==5.7.0
umap-learn==0.5.6
unicodedata2==15.0.0
urllib3==1.26.14
visdom==0.2.4
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.3
wheel==0.40.0
wheezy.template==3.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
xlrd==2.0.1
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
zict==2.2.0
zipp==3.15.0
zstandard==0.19.0
